backend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-CORS==4.0.0
google-cloud-firestore==2.15.0
google-cloud-storage==2.14.0
python-dotenv==1.0.0
gunicorn==21.2.0
